+++ test/Sentry.Tests/Internals/Http/HttpTransportTests.cs
@@ -31,8 +31,7 @@ namespace Sentry.Tests.Internals.Http
 
             var httpTransport = new HttpTransport(
                 new SentryOptions {Dsn = DsnSamples.ValidDsnWithSecret},
-                new HttpClient(httpHandler),
-                _ => { }
+                new HttpClient(httpHandler)
             );
 
             var envelope = Envelope.FromEvent(
@@ -73,8 +72,7 @@ namespace Sentry.Tests.Internals.Http
                     Debug = true,
                     DiagnosticLogger = logger
                 },
-                new HttpClient(httpHandler),
-                _ => { }
+                new HttpClient(httpHandler)
             );
 
             var envelope = Envelope.FromEvent(new SentryEvent());
@@ -113,8 +111,7 @@ namespace Sentry.Tests.Internals.Http
                     Debug = true,
                     DiagnosticLogger = logger
                 },
-                new HttpClient(httpHandler),
-                _ => { }
+                new HttpClient(httpHandler)
             );
 
             var envelope = Envelope.FromEvent(new SentryEvent());
@@ -137,7 +134,7 @@ namespace Sentry.Tests.Internals.Http
         public async Task SendEnvelopeAsync_ItemRateLimit_DropsItem()
         {
             // Arrange
-            using var httpHandler = new FakeHttpMessageHandler(
+            using var httpHandler = new FakeHttpClientHandler(
                 _ => SentryResponses.GetRateLimitResponse("1234:event, 897:transaction")
             );
 
@@ -146,8 +143,7 @@ namespace Sentry.Tests.Internals.Http
                 {
                     Dsn = DsnSamples.ValidDsnWithSecret
                 },
-                new HttpClient(httpHandler),
-                _ => { }
+                new HttpClient(httpHandler)
             );
 
             // First request always goes through
@@ -198,24 +194,22 @@ namespace Sentry.Tests.Internals.Http
         }
 
         [Fact]
-        public void CreateRequest_AuthHeader_Invoked()
+        public void CreateRequest_AuthHeader_IsSet()
         {
             // Arrange
-            var callbackInvoked = false;
-
             var httpTransport = new HttpTransport(
                 new SentryOptions {Dsn = DsnSamples.ValidDsnWithSecret},
-                new HttpClient(),
-                _ => callbackInvoked = true
+                new HttpClient()
             );
 
             var envelope = Envelope.FromEvent(new SentryEvent());
 
             // Act
-            httpTransport.CreateRequest(envelope);
+            using var request = httpTransport.CreateRequest(envelope);
+            var authHeader = request.Headers.GetValues("X-Sentry-Auth").FirstOrDefault();
 
             // Assert
-            callbackInvoked.Should().BeTrue();
+            authHeader.Should().NotBeNullOrWhiteSpace();
         }
 
         [Fact]
@@ -224,8 +218,7 @@ namespace Sentry.Tests.Internals.Http
             // Arrange
             var httpTransport = new HttpTransport(
                 new SentryOptions {Dsn = DsnSamples.ValidDsnWithSecret},
-                new HttpClient(),
-                _ => { }
+                new HttpClient()
             );
 
             var envelope = Envelope.FromEvent(new SentryEvent());
@@ -243,8 +236,7 @@ namespace Sentry.Tests.Internals.Http
             // Arrange
             var httpTransport = new HttpTransport(
                 new SentryOptions {Dsn = DsnSamples.ValidDsnWithSecret},
-                new HttpClient(),
-                _ => { }
+                new HttpClient()
             );
 
             var envelope = Envelope.FromEvent(new SentryEvent());
@@ -264,8 +256,7 @@ namespace Sentry.Tests.Internals.Http
             // Arrange
             var httpTransport = new HttpTransport(
                 new SentryOptions {Dsn = DsnSamples.ValidDsnWithSecret},
-                new HttpClient(),
-                _ => { }
+                new HttpClient()
             );
 
             var envelope = Envelope.FromEvent(new SentryEvent());