+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## unreleased
+## vNext
 
+* Implement envelope caching. (#576) @Tyrrrz
 * Add a list of .NET Frameworks installed when available. (#531) @lucas-zimerman
 * Parse Mono and IL2CPP stacktraces for Unity and Xamarin (#578) @bruno-garcia
 * Update TFMs and dependency min version (#580) @bruno-garcia